Fix username showing up in empty entries in Hall of Fame
When you start playing the game, the Hall of Fame is initially empty. When you complete the game, your name appears in the Hall of Fame, but not only in 1st place, but in 2nd and 3rd place too, with a time of 0:00:00. See screenshots below. Cause: the string containing the player name was not cleared when iterating over hall of fame entries.

Fixed by Krzysztof Nowaczyk, thank you! :heart: