Stop recording demo when a level is restarted
When I was recording a demo, and wanted to restart the level, I tried using <Esc>, "Restart level" instead of going back to the main menu. The demo buffer was not cleared, so all my inputs from the second attempt got glued together with the first attempt, making the demo nonsensical. I fixed this by stopping the demo recording when restarting a level.

Fixed by Krzysztof Nowaczyk, thank you! :heart: