Fix demo signatures never being appended
When recording a demo I found that if a signature is present in myspsig.txt it is never appended to the demo. That's because the file the game is trying to open is actually MYSPSIG.TXT, all uppercase, which caused a file open to fail. Another issue: when the signature is about to be written to the output file, the code tries to write to a wrong file handle. It tries to write to the signature file (which is open in read-only mode) instead of the demo file. One last bug: there was an "off by one" error when writing the signature to the output file and the last terminating 0xFF byte was not written to the file.

Fixed by Krzysztof Nowaczyk, thank you! :heart: