Fix bug with panel visibility and save states
When a state is saved when the panel was hidden, and that state is loaded with
the panel visible, the panel remains visible, but the playfield is
behind the panel. If you press Enter the panel remains visible and the
playfield is adjusted.
Likewise, if the panel is hidden and a state is loaded with the panel
visible, you need to press Enter to show it.

Fix that issue by setting the gCurrentPanelHeight variable when loading
a save state.